requirements.txt
# Core data science packages
pandas>=1.5.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.6.0
seaborn>=0.12.0
category_encoders>=2.5.0

# Machine learning and statistical analysis
scipy>=1.10.0
statsmodels>=0.14.0
xgboost>=1.7.0
lightgbm>=4.0.0

# Configuration and utilities  
pyyaml>=6.0
pathlib2>=2.3.7
joblib>=1.3.0

# Development and documentation
jupyter>=1.0.0
notebook>=6.5.0
ipykernel>=6.25.0

# Data validation and profiling
great_expectations>=0.17.0
pandas_profiling>=3.6.0

# Performance optimization
numba>=0.57.0
cython>=0.29.0

# Logging and monitoring
loguru>=0.7.0
tqdm>=4.65.0

# Testing (development)
pytest>=7.4.0
pytest-cov>=4.1.0
